Update ticket status when sending reply
This works around a couple of problems. With two different form buttons, we don't go through the form `onSubmit` anymore, so we can't use the clearOnSubmit feature of TextAreaInput. Instead, it now uses the latest message as a key, to generate a fresh input after a message is added successfully.

It also adds a guard against submitting an empty reply. This only works with the plain text version of the editor, rather than the HTML stored as `content`, since it will turn whitespace into `<p>`, `nbsp;`, etc. To get the plain text, we pass the editor onChange event back through InputHOC as a third argument to the provided onChange function... which we can then use to access the editor itself and get the plain text version of the content.