Refactor SpecialUsers and Setting, fix wikipedia_expert downgrade
This moves the setting of special users out of Setting and over to the SpecialUsers class instead, as Setting doesn't need to have any knowledge of the SpecialUsers implementation build atop it.

The if/else branches for handling `wikipedia_experts` as a group role is a little messy, but I can't think of a better solution at the moment.